Reimplement power/reset and host state with sysfs gpios
Remove the need for a permanently running daemon, as that was proving to
be a bit unreliable; there were also significant performance and power
consumption costs.

Note: the locator feature has been disabled as I don't personally use
it, and implementing it would very likely require a constantly running
process.  But I think that the implementation of the essential features
(host state; power and reset switches) should be kept separate.