Rewrite gpio_setup and fix setup errors
Includes a small delay between exporting a pin and trying to set its direction.
This is necessary to avoid permission errors on my original Model B Pi.

Additionally, if the pin has already been exported (by a previous execution of
bmc.sh), do not try to export it again and prevent the resulting error/warning.